feat : Implement generalizated StarRating Component for Movies, TV Shows ..etc.
 Description:
- Refactored the `StarRating` component to make it applicable for both movies and TV shows.
- Calculated the number of filled stars based on the provided rating, allowing for a maximum of 5 stars displayed.
- Utilized SVG icons for star representation, ensuring scalability and visual clarity.
- Styled the stars to dynamically change color based on the rating: filled stars appear in yellow, while unfilled stars are gray.
- Included a numeric rating display formatted as "rating/10" for user clarity.
- Ensured the component maintains a responsive layout, using flexbox for proper alignment.

This commit effectively summarizes the enhancements and generalization made to the `StarRating` component!
